-돌봄 예약금 적립 추상 인터페이스

+++ src/main/java/com/PetCare/service/Reservation/Reward/RewardService.java
@@ -1,6 +1,6 @@
 package com.PetCare.service.Reservation.Reward;
 
 public interface RewardService {
-    void addReward(long customerId, double amount); // 적립 메서드
-    double calculatorReward(double reservationAmount); // 적립금 계산 메서드
+    void addReward(long customerId, int amount); // 적립 메서드
+    int calculatorReward(int reservationAmount); // 적립금 계산 메서드
 }
